Move sold resource handling responsibility
